requirements.txt
# Core dependencies
kubernetes>=24.0.0
pyyaml>=6.0
jsonschema>=4.0.0
GitPython>=3.1.0
aiohttp>=3.8.0

# Real-time monitoring dependencies
channels>=4.0.0
channels-redis>=4.0.0
redis>=4.0.0

# Performance optimization dependencies
celery>=5.3.0
django-redis>=5.3.0
psutil>=5.9.0

# Development dependencies  
pytest>=7.0.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
pre-commit>=2.17.0

# GUI Testing dependencies (Playwright)
playwright>=1.40.0
pytest-playwright>=0.4.0
pytest-xvfb>=3.0.0

# Visual regression testing dependencies
pixelmatch>=0.3.0
Pillow>=9.0.0

# NetBox (for development)
# Note: In production, NetBox should already be installed
netbox>=3.4.0